Extract ProxySelector in its own class
